Remove unnecessary 32-bit code.
Since DTrace for Linux is only being developed for 64-bit platforms, various
pieces of conditional-controlled code can be removed because they only apply
for 32-bit DTrace implementations.  Note that this does not have anything to
do with (future) support for 32-bit applications runnings and being traced on
a 64-bit host.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>