dtrace: implement omni-present cyclics
This commit adds support for omni-present cyclics.  An omni-present
cyclic is one that fires on every CPU on the system, each at the same
frequency/interval.  They are implemented based on regular cyclics,
pinned to specific CPUs.  The implementation is such that hotplugging
CPUs is supported, i.e. when new CPUs come online, cyclics are started
on them for every omni-present cyclic that is currently active, and
when a CPU goes offline, the cyclics associated with that CPU (at least
those that are part of an omni-present cyclic) are removed.

Orabug: 18323501

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Jerry Snitselaar <jerry.snitselaar@oracle.com>