ctf: move the module->ctf_file info into a structure.
Rather than tracking the underlying ctf_file info in a module_to_ctf_file hash
mapping module name to the CTF file directly, introduce a structure to hold the
CTF file info in.

This has absolutely no effect on the output (nor even on the in-memory data
structures, since a structure with one member looks identical to the member on
its own) but is necessary for later commits that add further per-module state.

Orabug: 20229506

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Acked-by: Kris Van Hees <kris.van.hees@oracle.com>