dtrace: fix dtrace_helptrace_buffer memory leak
When the help tracing facility is enabled in DTrace, upon loading the
DTrace core module, a buffer was being allocated using vmalloc(), yet
is was never freed upon unloading of the dtrace module.  This caused a
leak of (by default) 64K with every load of the dtrace module.  This
commit ensures that the memory is freed.

The commit also fixes the problem that the help tracing facility
variables in DTrace were defined in two places.

Orabug: 20514336

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Acked-by: Nick Alcock <nick.alcock@oracle.com>