dtrace: convert from sdt_instr_t to asm_instr_t 2of2
The sdt_instr_t type was introduced as a generic type name for machine
instructions to be defined as a synonym for the arch-specific type.
It was named thusly because it was only used by the SDT support code.
However, other providers will need it also (e.g. FBT) and therefore it
is more appropriate to give it a name that is not tied to a specific
provider.

Orabug: 21220305

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Acked-by: Nick Alcock <nick.alcock@oracle.com>
Acked-by: Allen Pais <allen.pais@oracle.com>