kallsyms: make it possible to disable /proc/kallmodsyms
This allows you to build UEK without elfutils present at all, just like you
can with the upstream.

It requires a bit of fiddling about in kernel/kallsyms.c to make kallsyms
work with a NULL kallsyms_symbol_modules, and a bit of code motion to keep
the number of ifdefs down.

Verified that /proc/kallsyms has identical output (modulo only addresses,
dtrace-related symbols and, of course, the symbols ifdeffed out in this
patch) when the new CONFIG_KALLMODSYMS option is flipped.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Acked-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: David  Mc Lean <david.mclean@oracle.com>
Fwd-Ports: 4.10->4.11 4.12->4.13 4.13->4.14
Orabug: 21539840