dtrace: ensure SDT module probes work with NORX
When the CONFIG_DEBUG_SET_MODULE_RONX option is enabled in the kernel,
the executable code of the module is marked read-only prior to calling
the module notifier.  This causes the patching of NOPs in the probe
locations to fail with a system crash.

Because patching of SDT probe locations should happen upon module load
for DTrace-enabled kernels regardless of whether or not DTrace will be
used, it is both cleaner and more correct to make a direct call to the
code that handles the patching rather than using the notifier mechanism.
The new call takes place prior to marking the pages read-only, avoiding
the problem altogether.

Orabug: 21630297
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Acked-by: Nick Alcock <nick.alcock@oracle.com>