dtrace: ensure signal-handled is fired with correct signal
When sending a signal to a process that neither ignores nor handles
that signal, it is translated into a SIGKILL because the signal will
effectively cause the process to be terminated.  As a result, one
would see a signal-send probe for the original signal, and a
signal-handled probe for SIGKILL.

Since in these cases the original signal is retained within the
target task for exit reporting, it is possible to report the original
signal number in the signal-handled probe.  This commit accomplishes
that to ensure that accurate pairing for signal probes is possible.

Orabug: 22573604
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Acked-by: Nick Alcock <nick.alcock@oracle.com>