SPEC: dtrace-module specfile revamp.
Merge both the specfiles into one, and remove all conditionalizing for
older module versions: the specfile now relates only to the current
version, but applies to both OL6 and OL7.

The string @SOURCE_TARBALL@ must be replaced with the name of the source
tarball before building.  (The release autobuilders already do this.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>