dtrace: v4.13 changes to waitfd related code
In the code previous to kernel version 4.13, DTrace modifications added
a call in kernel/exit.c (named do_wait4) to perform an operation used by
fs/waitfd.c.  do_wait4 was constructed by taking the code called in a
SYSCALL_DEFINE and making a function out of it -- do_wait4() was then
used to implement the functionality of the syscall.

In 4.13, a similar thing has happened in the upstream kernel. That is,
kernel_wait4 has been created in kernel/exit.c and kernel_wait4() is
called inside SYSCALL_DEFINE4(wait4, ...
This patch borrows the new kernel_wait4() for our use by fs/waitfd.c,
and completes the functionality which operated in earlier kernels and
which was removed in the earlier commits:

  wait: add waitfd(), and a testcase for it
  wait: fix loss of error code from waitid() when info is provided
    (the above commit is completely removed now)
  wait: change waitfd() to use wait4(), not waitid(); reduce invasiveness

Signed-off-by: David  Mc Lean <david.mclean@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>