dtrace: fix retrieval of FBT entry probe arguments beyond the 5th
The retrieval of function arguments upon firing of an FBT entry probe
was not working correctly for arguments arguments that are stored on
the stack.  The skipping of frames (to access the stack frame where
the arguments are stored) was faulty because the number of frames to
skip changed for kernels beyond 4.1, and could result in a crash.

The number of frames to skip has been updated, and the implementation
has been made more robust for all probes that allow access to arguments.
If there is a valid stored set of registers available for the probe
location, it will be used directly.  If there is no such set available,
the kernel facility for stack frame unwinding will be used to safely
traverse the chained frames.

Orabug: 26974281
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>