ctf: pass the constituents of thin archives to dwarf2ctf
If we don't take special measures to avoid it, we end up passing vmlinux
(the whole thing) to dwarf2ctf for scanning.  Since dwarf2ctf uses input
filenames to determine what module things are part of, this leads to
every built-in module being considered to be part of the vmlinux module,
which is very far from ideal.  dwarf2ctf knows the mapping from object
filenames to module names, but this doesn't help if it's never told the
names of any of the modules (and computing them from the translation
unit names in vmlinux, while possible, would require an additional
scanning pass to determine which TUs were in which module, and we just
tore all that code out!)

Fix this by telling Makefile.modpost where to find the non-vmlinux
object files containing the built-in modules, and exploding the names of
any thin archives found among them into their constituent filenames
while building up the .ctf.filelist, just as we do when building up
modules_thick.builtin.  (These files do not serve the same purpose,
though the CTF filelist includes many .ko files for external modules
and also contains .o files for always-built-in kernel components, both
of which are omitted from modules_thick.builtin.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reported-by: Nicolas Droux <nicolas.droux@oracle.com>