dtrace: syscall provider does not fall back to slow syscall path
Syscalls can fall back from the fast path to the slow one if there
is more work to be handled or a full pt_regs structure is required.

The syscall wrapper installed by syscall provider breaks this fall
back. The result is that some syscalls like clone may corrupt some
of the userspace registers.

Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>