dtrace: handle failed allocations during probe providing
Solaris code was relying on never failing memory allocations. Thus
error code paths were not implemented. On Linux there is no guarantee
that memory allocation succeeds unless __GFP_NOFAIL is used; using
__GFP_NOFAIL unnecessarily is considered bad practice.

The correct fix is to add proper handling of memory allocation failures
to the DTrace framework.

This bug was uncovered by a toxic change in 4.13 kernels that caused
allocations to prematurely fail if a task has been killed. The toxic
commit has been backed out so it is not reproducible on latest stable
kernels.

Orabug: 27161612

Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>