dtrace: SDT probes cause panic when kaslr is enabled
SDT probes locations are using absolute addresses for the kernel
itself. This causes corruption of the .text section when the kernel
gets relocated to use a different base than the one it was linked to
(for example when kaslr is enabled).

This patch changes the way how SDT probes are encoded. After this
change the probe locations are going to be relocated during boot
with other kenel sections.

Orabug: 27211706

Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>