ctf: fix thinko preventing linking of out-of-tree modules when CTF is off
The CTF decoupling commit dropped a bunch of variable initializations
for the (degenerate) external-module no-CTF case because we would need
the same initializations for other cases, moving those initializations
further up and overriding them only when needed (in the CTF-enabled,
out-of-tree-module case).

Unfortunately I then forgot to move the containing ifdef CONFIG_CTF
down, leading to these variables being entirely unset in the out-
of-tree module case.  This causes linking of out-of-tree modules to fail
when CTF is off.

Thanks to Iain Barker and releng for tracking this down and reporting
it.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Chuck Anderson <chuck.anderson@oracle.com>
Orabug: 27215305