dtrace: Add dynamic debug support
DTrace contains various debugging output which is controled by
the CONFIG_DT_DEBUG. This is not enough to enable the actuall
debug output and modules needs to be rebuilt once more after
updating dtrace/dtrace_debug.h

This patch converts all debug messages into dynamic debug ones
(if it is supported by the kernel). This allows turning on and off
all/subset of messages as requried during runtime.

Orabug: 27396551

Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>