dtrace: allow SDT to be disabled even when DTrace is enabled
The include/linux/sdt.h file does not support SDT being disabled on
a system where DTrace itself is enabled.  That poses issues when
supporting DTrace on architectures where we do not have SDT probing
support available yet.

This commit makes the appropriate portions of sdt.h conditional on
whether SDT has been enabled.  This change also revealed that there
is a missing semicolon in dt_test_dev.c in a DTRACE_PROBE() call,
which is fixed in this commit also.

Orabug: 27457396
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tom Saeger <tom.saeger@oracle.com>