DTrace: initial ARM64 port of DTrace
This commit adds initial support for DTrace on ARM64.  It builds on
earlier commits that set the stage for ARM64 support to be added
(disabling SDT by itself, supporting architectures that pass up to
8 arguments in registers, adding waitfd syscall to the generic
syscall list if DTrace is enabled, and ARM64 support for SDT probe
discovery).

Note that there are other packages outside of the kernel that are
needed to build, test, and use DTrace on ARM64:
- libdtrace-ctf: Compact Type Format library.
- dtrace-user: DTrace consumer library and utility, and testsuite.

There are four components to the patch that cannot be split up further
because of the interdependence of the components:

- ARM64-specific kernel changes:
    + Marking ARM64 as an architecture that has DTrace support.
    + Adding a cpu_data(cpu) macro to access cpu_data for a specific
      cpu (consistent with other architectures in Linux).
    + Adding DIE_PAGE_FAULT to the list of supported die values,
      and add a die notifier call for page faults on ARM64 when DTrace
      is enabled (consistent with other architectures).
    + Make the sys_call_table be writable when DTrace is enabled
      (consistent with other architectures).

- ARM64 DTrace implementation:
    + Architecture specific code for the core kernel DTrace support
      code (in arch/arm64/include/asm/dtrace* and
      arch/arm64/kernel/dtrace*) for FBT, SDT, and systrace providers.
      The support for FBT and SDT is preliminary, to support the
      upcoming full implementation of those providers.
    + A placeholder FBT blacklist.
    + Architecture specific code for the DTrace core module
      (in arch/arm64/dtrace).

- x86_64 DTrace change:
    + dtrace_mod_pdata_init() and dtrace_mod_pdata_cleanup() are to
      be implemented by ever architecture that supports DTrace, so
      the function prototypes moved to dtrace_os.h instead.

For this initial version, DTrace on ARM64 supports the following
providers:
- profile
- systrace

The ARM64 support in the DTrace core is still limited, and support
for the full D language and actions will be added incrementally.

Orabug: 27438977
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Vincent Lim <vincent.lim@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tom Saeger <tom.saeger@oracle.com>